+++ Agil.Eccomerce.Api/Controllers/ProductsController.cs
@@ -64,28 +64,45 @@ namespace Agil.Eccomerce.Api.Controllers
             }
             else
             {
-                return BadRequest($"Exists a province with id{product.Id}");
+                return BadRequest($"Exists a product with id{product.Id}");
             }
             
         }
 
         // PUT api/<ProductsController>/productId
         [HttpPut("{id}")]
-        public void Update(int id, [FromBody] Product product)
+        public IActionResult Update(int id, [FromBody] Product product)
         {
-            var ProductToUpdate = _context.Producto.Single(p => p.Id == id);
-            _context.Producto.Remove(ProductToUpdate);
-            _context.Producto.Add(product);
-            _context.SaveChanges();
+            if (_context.Producto.Any(p => p.Id == id))
+            {
+                var ProductToUpdate = _context.Producto.Single(p => p.Id == id);
+                _context.Producto.Remove(ProductToUpdate);
+                _context.Producto.Add(product);
+                _context.SaveChanges();
+                return Ok();
+            }
+            else
+            {
+                return BadRequest($"Don't exist a product with id {id}");
+            }
         }
 
         // DELETE api/<ProductsController>/productId
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
-            var ProductToDelete = _context.Producto.Single(p => p.Id == id);
-            _context.Producto.Remove(ProductToDelete);
-            _context.SaveChanges();
+            if (_context.Producto.Any(p => p.Id == id))
+            {
+                var ProductToDelete = _context.Producto.Single(p => p.Id == id);
+                _context.Producto.Remove(ProductToDelete);
+                _context.SaveChanges();
+                return Ok();
+            }
+            else
+            {
+                return BadRequest($"Don't exist a product with id {id}");
+            }
+
         }
     }
 }
